PAV_TTE and PAV_TGE fixed, double build works
They had two separate issues;

1 - One input int type was unsigned instead of long,
which became an issue because it actually needed to
be signed. thus -33 -> HUGE NUMBER

2 - The padding_factor amendments were only imple-
mented on the PAV_TTI and PAV-combo kernels. This
meant that when double precision forced a split of
PAV and WAVG, it forced usage of PAV_TGE, which was
unaware that it should correct the euler matrix by
padding factor.

+++ src/gpu_utils/cuda_kernels/difference.cu
@@ -176,10 +176,10 @@ __global__ void cuda_kernel_wavg(
 //	{
 		for (unsigned pass = 0; pass < pass_num; pass ++)
 		{
-			s_wavgs_real[tid]  = 0.0f;
-			s_wavgs_imag[tid]  = 0.0f;
-			s_wdiff2s_parts[tid] = 0.0f;
-			Fweight = 0;
+			s_wavgs_real[tid]  = (FLOAT)0.0;
+			s_wavgs_imag[tid]  = (FLOAT)0.0;
+			s_wdiff2s_parts[tid] = (FLOAT)0.0;
+			Fweight = (FLOAT)0.0;
 
 			pixel = pass * BLOCK_SIZE + tid;
 			s_Minvsigma2s[tid]=g_Minvsigma2s[pixel];
